LinkInfo: retry HTML parsing with raw content
HTML parsing with lxml fails if the string has already been decoded to
unicode but the HTML contains an XML encoding declaration.  Since not
all pages have this feature, and we'd like to preserve requests'
encoding detection where possible, our best bet is to try again with
just the raw bytes and let lxml use the declared encoding in the
document.