Better UX
This change makes !set always succeed in setting the whois for this channel, but it will now never over-write the global value if one exists.
It also means that !unset behaves similarly, it always succeeds in unsetting whichever whois was set for the channel, including removing the global one should it exist.
I think more work needs to be done to decide whether that last decision is a good one: removing/setting global data on accident is bad UX but it's better than what we currently have, and better than unexpectedly being unable to set the whois.